Fix to submitting votes causing a crash
