+++ thavalon-web/src/components/proposalComponents.tsx
@@ -1,5 +1,7 @@
 import React from "react";
 import { Button, Form, ToggleButtonGroup, ToggleButton } from "react-bootstrap";
+import { WebSocketManager, WebSocketProp } from "./communication";
+import { OutgoingMessageTypes } from "../Core/commConstants";
 
 //#region Interfaces
 interface ProposalSelectionFormProps {
@@ -11,9 +13,111 @@ interface ProposalSelectionFormProps {
 interface ProposalSelectionFormState {
     proposedPlayers: string[];
 }
+
+interface ProposalUIProps extends WebSocketProp {
+    proposer: string;
+    isProposing: boolean;
+    proposal: string[];
+    numOnProposal: number;
+    playerOrder: string[];
+}
+
+interface TentativeProposalMessage {
+    type: OutgoingMessageTypes.SubmitProposal;
+    proposal: string[]
+}
+
+interface OutgoingMoveToVoteMessage {
+    type: OutgoingMessageTypes.MoveToVote
+    proposal: string[];
+}
 //#endregion
 
-export class ProposalSelectionForm extends React.Component<ProposalSelectionFormProps, ProposalSelectionFormState> {
+export class ProposalUI extends React.Component<ProposalUIProps, {proposal: string[]}> {
+    private _connection: WebSocketManager;
+    constructor(props: ProposalUIProps) {
+        super(props);
+        this._connection = props.webSocket;
+        this.state = {proposal: []};
+    }
+
+    render(): JSX.Element {
+        if (this.props.isProposing) {
+            return this.createProposerUI();
+        } else {
+            return this.createOtherProposerUI();
+        }
+    }
+
+    private createProposerUI(): JSX.Element {
+        let currentProposal: JSX.Element | undefined = undefined
+        if (this.state.proposal.length !== 0) {
+            const playerList = this.state.proposal.map((player) => {
+                return <li key={player}>{player}</li>;
+            });
+            currentProposal = <ul>{playerList}</ul>;
+        }
+        return (
+            <span>
+                {currentProposal}
+                <ProposalSelectionForm 
+                    callback={(proposal: string[]) => {
+                        this.sendTentativeProposal(proposal);
+                    }}
+                    numOnProposal={this.props.numOnProposal}
+                    playerOrder={this.props.playerOrder} />
+                <br />
+                <Button 
+                    type="button" 
+                    onClick={this.moveToVote.bind(this)}>
+                    Move to Vote
+                </Button>
+            </span>
+        );
+    }
+
+    private createOtherProposerUI(): JSX.Element {
+        const proposal = this.props.proposal;
+        let proposalInfo: JSX.Element
+        if (proposal.length === 0) {
+            proposalInfo = (
+                <span>
+                    Please wait while {this.props.proposer} proposes a team.
+                </span>
+            );
+        } else {
+            proposalInfo = this.formatProposalList();
+        }
+        return proposalInfo;
+    }
+
+    private formatProposalList(): JSX.Element {
+        const proposal = this.props.proposal;
+        const proposalList = proposal.map((player) => {
+            return <li key={player}>{player}</li>
+        });
+        return (<ul>{proposalList}</ul>);
+    }
+
+    private sendTentativeProposal(proposal: string[]): void {
+        const message: TentativeProposalMessage = {
+            type: OutgoingMessageTypes.SubmitProposal,
+            proposal: proposal
+        };
+        this._connection.send(message);
+        this.setState({proposal: proposal});
+    }
+
+    private moveToVote(): void {
+        const message: OutgoingMoveToVoteMessage = {
+            type: OutgoingMessageTypes.MoveToVote,
+            proposal: this.state.proposal
+        };
+        this._connection.send(message);
+    }
+}
+
+class ProposalSelectionForm extends React.Component<ProposalSelectionFormProps, ProposalSelectionFormState> {
     constructor(props: ProposalSelectionFormProps) {
         super(props);
         this.state = {proposedPlayers: []}
