+++ thavalon/consumers.py
@@ -295,7 +295,7 @@ class GameConsumer(WebsocketConsumer):
         self.send(response.serialize())
 
     def broadcast_moving_to_vote(self, proposal):
-        self.game.set_proposal(proposal)
+        self.game.set_proposal(proposal.get("proposal"))
         vote_info_event = {"type": "send_vote_info", "proposal": proposal}
         async_to_sync(self.channel_layer.group_send)(self.lobby_group_name, vote_info_event)
